Send url to cucumber tests (#29)
* Add cucumber tests

Adding dummy cucumber tests, with a task that execute it

* Exposing port without port forward

I do not want to run it with port forward. I want to make it easier to run the server. By the moment I made it by exposing the ports on the cluster.

* Run the cucumber tests with a configuration.

We can now run the test on a diferent ulr than kind. So you are able to run the test with a main application outside th kind cluster